fix(daemon): close rebuild-index safety gaps from CodeRabbit review
Problem: PR #3076 (daemon-coordinated online index rebuilds) shipped with 8
CodeRabbit findings never addressed, tracked as debt in polylogue-ogn1
because the merge task was scoped to a conflict-free rebase, not review.

What changed, verified against current source (not blindly reapplied):

1. HTTP handler must reject max_blob_mb without raw_ids/only_missing before
   promoting -- ALREADY FIXED prior to merge (validate_rebuild_index_request
   already enforced this in the merged commit 1c2a07f9b; confirmed via
   `git show 1c2a07f9b:polylogue/maintenance/rebuild_index.py`). No change.

2/5. write_coordinator.py's bridge held a fixed ~30s run_sync timeout while
   the CLI/HTTP rebuild contract allows up to 600s. Added
   DaemonWriteThreadBridge.run_sync_with_timeout (a per-call override,
   scoped to this need rather than a general timeout-override framework);
   run_sync now delegates to it with the bridge's own default. The
   rebuild-index HTTP route calls it with a 600s budget matching the CLI's
   own urlopen(..., timeout=600).

3. CLI --daemon-url default read POLYLOGUE_DAEMON_URL directly via
   os.environ.get, bypassing the 5-layer config precedence chain every
   other daemon-URL consumer goes through. Now resolves via
   load_polylogue_config().daemon_url, matching status.py's own
   _default_daemon_url pattern.

4. HTTP handler fell back to executing the rebuild directly when
   server.write_bridge was missing, instead of failing closed. A real
   DaemonAPIHTTPServer always installs write_bridge in __init__, so this
   was unreachable in production, but a route that CAN execute an
   authority-promoting archive write outside the sole-writer coordinator is
   a bypass of the daemon's single-writer invariant, not a safe fallback.
   Now returns 503 write_coordinator_unavailable instead.

6. missing_index_raw_ids returned [] whenever index.db did not exist yet,
   making --only-missing rebuild nothing on a fresh archive or right after
   `ops reset --index` -- exactly the case it exists to handle. Now falls
   back to the full source set (every row is missing by definition).

7/9. Partial-selection-cannot-promote and empty-source-count-0 guards --
   ALREADY enforced inside the shared rebuild_index_from_source service
   itself (validate_rebuild_index_request called first thing; raw_count==0
   short-circuit before any replay), not merely the CLI. Confirmed present
   in the original merged commit. No change.

8. rebuild_index.py (substrate) imported the CLI's private
   _archive_readiness_status, inverting this repo's documented
   surfaces-may-not-import-substrate rule. Extracted the readiness-surface
   computation (archive_readiness_status + its _archive_readiness_counts/
   _action_readiness_counts/_archive_status_surfaces helpers) into
   polylogue/storage/archive_readiness.py; status.py now delegates to the
   same shared implementation instead of owning the only copy.

10. test_rebuild_index_route_uses_the_bridge_run_sync_writer_path replaced
   _handle_rebuild_index wholesale with its own body calling bridge.run_sync
   -- it only proved the test's own stand-in worked, never the real
   production handler. Rewritten to drive the real _do_post_impl dispatch
   and the real _handle_rebuild_index implementation end to end.

Solution: polylogue/daemon/write_coordinator.py (run_sync_with_timeout),
polylogue/daemon/http.py (fail-closed + 600s timeout on rebuild-index),
polylogue/cli/commands/maintenance/_rebuild_index.py (config-resolved
daemon URL), polylogue/maintenance/rebuild_index.py (missing_index_raw_ids
fix), polylogue/storage/archive_readiness.py (readiness extraction),
polylogue/cli/commands/status.py (delegates instead of owning). New/updated
tests prove each fix: a rejected unsafe fallback (test_maintenance_endpoints
.py), a real per-call timeout override (test_write_coordinator.py), a
fresh-index --only-missing selection (test_rebuild_index_selection.py), and
a real-dispatch rebuild-index route proof (test_http_write_coordination.py).

Verification:
- mypy polylogue (strict): Success: no issues found in 1082 source files
- ruff check . / ruff format --check .: All checks passed / all formatted
- devtools verify --quick: exit_code 0 (includes degrade-loudly, now
  allowlisted for the two extracted typed-signal except-handlers)
- devtools render all --check: no "out of sync" lines, exit 0
- devtools test across the affected area (status/rebuild_index/daemon http/
  write_coordinator/maintenance CLI): 264 passed, 1 pre-existing unrelated
  failure (test_archive_facade_route_catalog_covers_public_async_facade,
  confirmed failing identically on origin/master before this change via
  `git stash`)

Ref polylogue-ogn1
